requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
pytest==8.2.0
python-gitlab==4.4.0
requests==2.31.0
requests-mock==1.12.1
requests-toolbelt==1.0.0
types-requests==2.31.0.20240406
urllib3==2.2.1
